Finish optional SDL3 support: MSVC configs, makefiles, CI and packaging
- MSVC: new Debug-SDL3/Release-SDL3 configurations using a derived
  BaseConfiguration property; SDL include/lib dirs, import libs, USE_SDL3
  define and post-build DLL copy are parametrized. Helper projects map to
  the plain Debug/Release configs in the solution.
- GNU makefiles: USE_SDL3=1 knob on Linux, MinGW x64 (bundled SDL3) and
  MSYS2 clangarm64 (pkg-config). in_sdl*/snd_sdl* backends are now
  self-guarding and always compiled; selection is by define only.
- main_sdl.c includes SDL3/SDL_main.h (header-only WinMain shim, SDL3 has
  no SDL3main library).
- CI: SDL3 legs for MSVC, mingw64, clangarm64 and macOS; AppImage docker
  image builds SDL3 from source and compile-checks the SDL3 backend.
- Fix SDL3 backend issues found by MSVC /W4: TASK_TIMEOUT_INFINITE
  signedness, invalid SDL_JoystickID is 0 in SDL3, gamepad events use
  gdevice (cdevice is the camera event in SDL3).
- snd_sdl3.c: use SDL_Pause/ResumeAudioStreamDevice, log the device name.

+++ Quake/tasks.h
@@ -80,7 +80,9 @@ void TestTasks_f (void);
 #endif
 
 #ifdef USE_SDL3
-#define TASK_TIMEOUT_INFINITE -1
+// SDL3 timeouts are Sint32 milliseconds with -1 meaning infinite; Task_Join
+// takes uint32_t, so this converts back to -1 when passed through to SDL.
+#define TASK_TIMEOUT_INFINITE ((uint32_t)-1)
 #else
 #define TASK_TIMEOUT_INFINITE SDL_MUTEX_MAXWAIT
 #endif
